удаляет лишнюю логику при обновлении вопроса

+++ app/models/common/category_progress_info.py
@@ -13,21 +13,17 @@ if TYPE_CHECKING:
 class CategoryProgressInfo(Base):
     __tablename__ = "categories_progress_info"
 
-    user_id: Mapped[int] = mapped_column(
-        ForeignKey("users.id", ondelete="CASCADE"), nullable=False, primary_key=True
-    )
+    user_id: Mapped[int] = mapped_column(ForeignKey("users.id", ondelete="CASCADE"), nullable=False, primary_key=True)
     category_id: Mapped[int] = mapped_column(
         ForeignKey("categories.id", ondelete="CASCADE"),
         nullable=False,
         primary_key=True,
     )
-    level_id: Mapped[int] = mapped_column(
-        ForeignKey("levels.id", ondelete="CASCADE"), nullable=False, primary_key=True
-    )
+    level_id: Mapped[int] = mapped_column(ForeignKey("levels.id", ondelete="CASCADE"), nullable=False, primary_key=True)
     score: Mapped[int] = mapped_column(default=0, server_default="0")
 
     @property
-    def level_value(self) -> int:
+    def level_value(self) -> float:
         return self.level.value
 
     category: Mapped["Category"] = relationship("Category")
